// pronti-frontend-main/front/src/app/modulos/refinanciamiento/form/form-refinanciamiento.enum.ts
export enum FormRefinanciamientoEnum {
  // nombre = 'nombre',
  totalCuotas = 'totalCuotas',
  totalCuotasPagadas = 'totalCuotasPagadas',
  totalCuotasMora = 'totalCuotasMora',
  totalCuotasPagadasRefinanciamiento = 'totalCuotasPagadasRefinanciamiento',
  totalCuotasFaltantesRefinanciamiento = 'totalCuotasFaltantesRefinanciamiento',
  cuotasRestantesSinMora = 'cuotasRestantesSinMora',
  valorCuota = 'valorCuota',
  valorPendientePago = 'valorPendientePago',
  valorAgregarseCuota = 'valorAgregarseCuota',
  fechaRefinanciamiento = 'fechaRefinanciamiento',

}